### 2022-06-03 #4449 @editorialbot
[PRE REVIEW]: tell: a Python package to model future total electricity loads
**Submitting author:** @mcgrathc (<a href="http://orcid.org/0000-0002-8808-8312">Casey McGrath</a>)
**Repository:** https://github.com/IMMM-SFA/tell
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.0
