### 2022-06-03 #4449 @danielskatz
@mcgrathc - please remove the period from the title in your paper.md file, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
