+++ paper/paper.bib
@@ -24,8 +24,10 @@
 
 @article{binsted2022,
   title={GCAM-USA v5.3_water_dispatch: Integrated modeling of subnational U.S. energy, water, and land systems within a global framework},
-  author={Binsted, M. and Coauthors},
-  journal={Geoscientific Model Development Discussions},
+  author={Binsted, M., Iyer, G., Patel, P., Graham, N. T., Ou, Y., Khan, Z., Kholod, N., Narayan, K., Hejazi, M., Kim, S., Calvin, K., and Wise, M.},
+  journal={Geoscientific Model Development},
+  volume={15(6)},
+  pages={2533-2559},
   year={2022},
   publisher={EGU},
   doi={10.5194/gmd-15-2533-2022},
