+++ paper/paper.bib
@@ -23,7 +23,7 @@
 }
 
 @article{binsted2022,
-  title={GCAM-USA v5.3_water_dispatch: Integrated modeling of subnational U.S. energy, water, and land systems within a global framework},
+  title={GCAM-USA v5.3_water_dispatch: Integrated modeling of subnational {U.S.} energy, water, and land systems within a global framework},
   author={Binsted, M., Iyer, G., Patel, P., Graham, N. T., Ou, Y., Khan, Z., Kholod, N., Narayan, K., Hejazi, M., Kim, S., Calvin, K., and Wise, M.},
   journal={Geoscientific Model Development},
   volume={15(6)},
