+++ tell/install_raw_data.py
@@ -19,7 +19,8 @@ class InstallRawData:
 
     # URL for DOI-minted TELL raw data data package hosted on Zenodo:
     DATA_VERSION_URLS = {'0.0.1': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1',
-                         '0.1.0': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1'}
+                         '0.1.0': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1',
+                         '0.1.1': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1'}
 
     def __init__(self, data_dir=None):
 
