### 2022-06-03 #4449 @danielskatz
I think when you edit a comment to make it a command, editorialbot doesn't see it
