### 2022-06-08 #4449 @fraukewiese
@chrklemm , @jasondegraw , @euronion : Would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
The submission under review here is [tell: a Python package to model future total electricity loads #4449](https://github.com/IMMM-SFA/tell)
