### 2022-06-14 #4472 @editorialbot
[REVIEW]: tell: a Python package to model future total electricity loads.
**Submitting author:** @mcgrathc (<a href="http://orcid.org/0000-0002-8808-8312">Casey McGrath</a>)
**Repository:** https://github.com/IMMM-SFA/tell
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
