### 2022-07-04 IMMM-SFA/tell#31 @euronion
[JOSS] LICENSE and DISCLAIMER
The repository contains a LICENSE and DISCLAIMER file which seem to be partially overlapping.

Is the DISCLAIMER obsolete or is it intended to extend the LICENSE file? I'm not a legal expert, but in the latter case i presume the LICENSE would no longer be BSD-2 but a modified one.

Can you check the DISCLAIMER file against the LICENSE and remove it if obsolete?

xref https://github.com/openjournals/joss-reviews/issues/4472
