### 2022-07-04 IMMM-SFA/tell#32 @euronion
[JOSS] Question regarding paper authors
This question relates to the paper authors:

The list include 9 people, whereas contributions to this repository have been made by only 5 of them.
As such it is not obvious to me whether all authors fulfill the [authorship criteria](https://joss.readthedocs.io/en/latest/submitting.html#authorship) by JOSS.

Maybe you can briefly state the type of involvement of the other authors in the software?

xref https://github.com/openjournals/joss-reviews/issues/4472
