### 2022-07-04 IMMM-SFA/tell#33 @euronion
[JOSS] Expand installation instructions
The installation instructions here:
*  https://immm-sfa.github.io/tell/README.html#install-tell
* https://immm-sfa.github.io/tell/tell_quickstarter.html#install-tell

are a bit meager. Consider adding some short comments with the following information:

* It is a Python package which requires a Python and pip installed
* Which Python versions are supported
* Whether there are platform limitations (Windows, Linux, MacOS) or whether the software has been tested to work on all plattforms

Installation did work for me (WSL2 on Windows 10 with Python=3.10.5) without problems.

xref: https://github.com/openjournals/joss-reviews/issues/4472
