### 2022-07-04 IMMM-SFA/tell#35 @euronion
[JOSS] Low test code coverage
The test coverage for the code is low (~31.40% according to CodeCov): https://app.codecov.io/gh/IMMM-SFA/tell ).

You might want to consider expanding the test coverage and include some more in-detail testing than e.g. the high level functionality tests in [`test_mlp_predict.py`](https://github.com/IMMM-SFA/tell/blob/main/tell/tests/test_mlp_predict.py).

xref https://github.com/openjournals/joss-reviews/issues/4472
