### 2022-07-04 IMMM-SFA/tell#36 @euronion
[JOSS] Comments regarding documentation
I have a couple of comments and suggestions after working through the documentation:

1. Installation

> tell is based on open-source publicly accessible data.
(https://immm-sfa.github.io/tell/tell_quickstarter.html#install-the-package-of-data-underpinning-tell)

The term "open data" rather than "open source" is more appropriate, cf. e.g. [this post](https://forum.openmod.org/t/definitions-for-open/1407)

2. API reference

From the API reference on the website I was unable to find information on `tell.train` as mentioned in the quickstarter notebook. Maybe one or more modules are not properly included in the API reference?

3. Configuration

The quickstarter mentions default settings for MLP training as stored in the GitHub repository. I think these settings should be linked (to the respective file) or included in the documentation.
* Consider linking the file or including the file contents in the documentation
It might be potentially interesting for users to look up and change these settings in their local setup; especially if these settings might change between versions:
* Consider describing the local availability of the settings file in the Python installation location ( `tell.__file__` folder) so show the user how to check the values used by their version

4. Spatial resolution

For temporal resolution the documentation states hourly resolution. For spatial resolution, the user has to read quite far to find:
* This package is only inteded for the CONUS region, a subset of USA
* What the adequate spatial resolution is ("Work at a spatial resolution adequate for input to a unit commitment/economic dispatch (UC/ED) model.").

I suggest both aspects, as they are properties of major relevance to interested potential users, should be placed as prominently as the information on the hourly resolution.

I also suggest you highlight the aspect of geographical scope and resolution in the beginning of the JOSS paper.

5. Target audience

Consider specififying the target audience and users of your package in a popular location like [here](https://immm-sfa.github.io/tell/index.html)

xref https://github.com/openjournals/joss-reviews/issues/4472
