### 2022-07-04 IMMM-SFA/tell#37 @euronion
[JOSS] Comments and suggestions on paper
* The model and data used seems to be specific for the USA / the 54 BAs in the CONUS area of the USA. This should be highlighted more strongly in the paper and in the documentation. I also recommend considering adding this limitation (US-only model) prominently to the paper title.

* In the paper it is stated that
  > This model accounts for a wide range of other factors affecting electricity demand, including technology change in the building sector, energy prices, and demand elasticities.

  Based on my understanding of the paper and the description in the documentation these capabilities of the model stem from coupling it to other models like GCAM-USA. The capabilities of the model do not seem to be inherent (e.g. detailed technical-bottom-up modelling). I suggest this is made more clear in the paper that to use these capabilities, external assumptions and results have to be provided.

* Public vs open data (Line 47f):
  > tell is based entirely on publicly available data ...

    The example data is uploaded to Zenodo under CC-BY-4.0 license. If the data is not only available publicly, but also accessible under an open license, consider emphasising this here (e.g. "tell is based entierly on publicly available, open data ...")

* The software also provides easy access to statistical validation results between the forecasts and observed data (cf. example notebooks). I consider that an aspect worthwhile to be mentioned in the paper
* Consider adding the links in line 83-87 as markup links to the respective site, e.g.  [GitHub repository](https://github.com/IMMM-SFA/tell)

**References**
* A reference to the IM3's future climate and population scenarios (line 67f) should be added
* Your model relies on sklearn. Consider citing the sklearn software package: https://scikit-learn.org/stable/about.html#citing-scikit-learn

**Misc**
* Rewrite "open-source" as "open source"

xref https://github.com/openjournals/joss-reviews/issues/4472
