### 2022-07-04 #4472 @euronion
Modelled load time-series and projections based on different development scenarios and at high spatial and temporal resolution are essential for electricity system modelling. `tell` does offer this functionality for a limited scope (CONUS area in the US) based on a neural network approach.

I think what this package covers is a valuable contribution and I see the substantial scholarly effort given (the first commit in the repository dates back to 2020 when I checked earlier today). As such I also think it is suited for JOSS.

There are a few issues I have raised in the repository regarding usage, documentation and the paper draft which are now open in the `tell` GitHub repository before I can continue with the checklist.
