### 2022-07-06 IMMM-SFA/tell#32 @mcgrathc
@euronion, there are 5 authors that have made direct contributions to the repository:

- Casey R. McGrath
- Casey D. Burleyson
- Aowabin Rahman
- Travis Thurber
- Chris R. Vernon

There are 3 authors that did not directly contribute to the repository, but were critical in developing this software:

- Jennie S. Rice: Substantial contributions to the conception and development of the software. Significantly contributed to paper revisions. Final approval of the version to be published.
- Nathalie Voisin:  Substantial contributions to the development of the software. Provided interpretation of model results. Significantly contributed to paper revisions.
- Zarrar Khan: Substantial contributions to the development of the software. Provided critical feedback when making decisions on model inputs.

This adds to 8 authors in total, which we believe all made significant scholarly contributions to the software.
