### 2022-07-11 IMMM-SFA/tell#31 @mcgrathc
The license and disclaimer are both required by our institution (PNNL) as part of their open-source software protocols. We cannot modify them. We’ve used this combination in other software packages (e.g., [CERF](https://github.com/IMMM-SFA/cerf)) that have gone through review and have been previously published in JOSS.
