### 2022-07-18 #4472 @robbiemorrison
## Legal context of reviewed software

@euronion asked me to review an issue they raised here: https://github.com/IMMM-SFA/tell/issues/31

And here is my response.  I apologize in advance for the complexities involved.

The matter concerns a disclaimer provided in addition to a recognized open source license and any adverse interactions that might arise.

The particular issue was recently discussed on an open source law forum operating under the [Chatham House Rule](https://www.wikipedia.org/wiki/Chatham_House_Rule).  Four people contributed to that debate.  As would be expected, there was some variation in the views expressed.

All four agreed that the Battelle Memorial Institute disclaimer (hereafter BMI‑D) document does not impose further obligations on the user, grant any additional rights, or remove any rights granted to the user under the BSD‑2‑Clause license document.  In that regard, the **two documents are legally compatible**.

The main point of divergence was the **scope of the BMI‑D document**.  One person argued that the disclaimer would only apply to code contributed under Battelle Memorial Institute held copyright and not to other contributors.  Two people argued that, under the terms of the BSD‑2‑Clause, the BMI‑D document could be omitted on redistribution of the codebase.  That latter view is not entirely clear because the BSD‑2‑Clause mentions "the following disclaimer" should be retained without being fully explicit as to whether that phrasing would cover third‑party statements or not.  The two people raising that broader possibility agreed that their underlying legal claim was weak.

One person suggested that the codebase should not be described as "open source" because the BMI‑D document had not been previously approved by the [Open Source Initiative](https://www.wikipedia.org/wiki/Open_Source_Initiative) or the [Free Software Foundation](https://www.wikipedia.org/wiki/Free_Software_Foundation).  I&nbsp;guess that viewpoint is quite technical.  Another person opined that the codebase should not be described as "licensed under BSD‑2‑Clause" due to the additional disclaimer.

Taking a&nbsp;**liberal view**: the BSD‑2‑Clause and the BMI‑D documents are legally compatible, the scope of the BMI‑D applies only to code contributed by Battelle Memorial Institute solely in the context of this repository and does not apply to redistributions of the codebase, and the project does broadly warrant the descriptor "open source" even if redistribution is required to fully satisfy that claim.

Worth observing too that straying outside **accepted licensing practice** often adds to legal uncertainty and may not necessarily resolve the issues being addressed.  As this example shows.

The person writing this note has no legal training and this summary should not be seen as providing advice.

Copyright (c) 2022 Robbie Morrison &lt;<robbie.morrison@posteo.de>&gt;. This posing is licensed under a Creative Commons Attribution 4.0 International (CC‑BY‑4.0) License.
