### 2022-07-27 8ad86eb
Create CONTRIBUTING.md
add contributing md for bugs and issues
