+++ CONTRIBUTING.md
@@ -1,0 +1,23 @@
+[Issues]: https://github.com/IMMM-SFA/tell/issues
+[new Issue]: https://github.com/IMMM-SFA/tell/issues/new/choose
+[new Pull Request]: https://github.com/IMMM-SFA/tell/compare
+[IM3]: https://im3.pnnl.gov/
+
+## How to contribute
+___
+
+__I found a bug!__
+* Nice sleuthing!
+* Check if the bug has already been reported by searching the existing GitHub [Issues]. If you find a match, consider adding additional details to the existing ticket.
+* Open a [new Issue], being sure to include a clear title and description along with as much detail as possible; code samples or log messages demonstrating the bug are quite helpful.
+
+__I fixed a bug!__
+* First of all, thanks!
+* Open a [new Pull Request] with the fix. Ensure the description clearly outlines the bug and the solution. Include the Issue number if applicable.
+
+__I created a new feature!__
+* You're the best!
+* Consider opening a [new Issue] to describe use cases for the new feature. This will offer a platform for discussion and critique.
+* Then, open a [new Pull Request] with clear documentation of the methodology. Be sure to include new unit tests if appropriate.
+
+The [IM3] team truly appreciates and encourages community involvement. We're all in this together! 
