### 2022-07-27 d645c43
fix formatting in contribution guidelines
