+++ docs/source/contributing.rst
@@ -9,28 +9,30 @@ ______________
 
 If you find a bug or would like to contribute a new feature, please `open an issue <https://github.com/IMMM-SFA/tell/issues>`_.
 
-[Issues]: https://github.com/IMMM-SFA/tell/issues
-[new Issue]: https://github.com/IMMM-SFA/tell/issues/new/choose
-[new Pull Request]: https://github.com/IMMM-SFA/tell/compare
-[IM3]: https://im3.pnnl.gov/
 
 Contribution workflow (bug or feature)
-______________
+______________________________________
 
-__I found a bug!__
-* Check if the bug has already been reported by searching the existing GitHub [Issues]. If you find a match, consider adding additional details to the existing ticket.
-* Open a [new Issue], being sure to include a clear title and description along with as much detail as possible; code samples or log messages demonstrating the bug are quite helpful.
+1. I found a bug!
+* Check if the bug has already been reported by searching the existing GitHub
+`Issues <https://github.com/IMMM-SFA/tell/issues>`_. If you find a match, consider adding additional details to the
+existing ticket.
+* Open a `new Issue <https://github.com/IMMM-SFA/tell/issues/new/choose>`_, being sure to include a clear title and
+description along with as much detail as possible; code samples or log messages demonstrating the bug are quite helpful.
 
-__I fixed a bug!__
-* Open a [new Pull Request] with the fix. Ensure the description clearly outlines the bug and the solution. Include the Issue number if applicable.
+2. I fixed a bug!
+* Open a `new Pull Request <https://github.com/IMMM-SFA/tell/compare>`_ with the fix. Ensure the description clearly
+outlines the bug and the solution. Include the Issue number if applicable.
 
-__I created a new feature!__
-* Consider opening a [new Issue] to describe use cases for the new feature. This will offer a platform for discussion and critique.
-* Then, open a [new Pull Request] with clear documentation of the methodology. Be sure to include new unit tests if appropriate.
+3. I created a new feature!
+* Consider opening a `new Issue <https://github.com/IMMM-SFA/tell/issues/new/choose>`_ to describe use cases for the
+new feature. This will offer a platform for discussion and critique.
+* Then, open a `new Pull Request <https://github.com/IMMM-SFA/tell/compare>`_ with clear documentation of the
+methodology. Be sure to include new unit tests if appropriate.
 
 
 Contribution workflow (pull request)
-_____________________
+____________________________________
 
 The following is the recommended workflow for contributing to **tell**:
 
