### 2022-08-02 7da61a1
add comments to installation instructions
