+++ docs/source/tell_quickstarter.rst
@@ -55,7 +55,9 @@ not to save these plots by setting the ``save_images`` and
 -------------------
 
 ``tell`` is available via GitHub repository by using the pip install
-functionality.
+functionality. ``tell`` requires a Python version between 3.8 and 4.0 as
+well as a pip install to import the package. ``tell`` has been tested on
+Windows and Mac platforms.
 
 .. code:: ipython3
 
