### 2022-08-03 IMMM-SFA/tell#33 @mcgrathc
The following was added to the installation instructions in the README and quick-starter:

``tell` is available via GitHub repository by using the pip install functionality. `tell` requires a Python version between 3.8 and 4.0 as well as a pip install to import the package. `tell` has been tested on
Windows and Mac platforms.`

The changes can be found in this PR: https://github.com/IMMM-SFA/tell/pull/41
