### 2022-08-03 8872325
add notebook note on statistical validation results
