### 2022-08-04 5f8c2c2
add paper/documentation edits
