+++ paper/paper.bib
@@ -129,3 +129,12 @@
  pages={2825--2830},
  year={2011}
 }
+
+@article{wrf,
+ title={A Description of the Advanced Research WRF Version 4},
+ author={Skamarock, W. C., Klemp, J. B.  Dudhia, J., Gill, D. O., Liu, Z., Berner, J.,
+        Wang, W., Powers, J.G., Duda, M. G. , Barker, D. M. and Huang  X.-Y.},
+ journal={CAR Tech. Note NCAR/TN-556+STR},
+ pages={145},
+ year={2019}
+}
