### 2022-08-05 ea7ac29
open-source replaced by open data in quickstarter install info
