### 2022-08-05 IMMM-SFA/tell#37 @mcgrathc
Please see this PR for the changes: https://github.com/IMMM-SFA/tell/pull/41

Public vs open data (Line 47f):
tell is based entirely on publicly available data ...
The example data is uploaded to Zenodo under CC-BY-4.0 license. If the data is not only available publicly, but also accessible under an open license, consider emphasising this here (e.g. "tell is based entierly on publicly available, open data ..."): **adopted**

The software also provides easy access to statistical validation results between the forecasts and observed data (cf. example notebooks). I consider that an aspect worthwhile to be mentioned in the paper: **adopted**

Consider adding the links in line 83-87 as markup links to the respective site, e.g. [GitHub repository](https://github.com/IMMM-SFA/tell): **adopted**

References
A reference to the IM3's future climate and population scenarios (line 67f) should be added
Your model relies on sklearn. Consider citing the sklearn software package: https://scikit-learn.org/stable/about.html#citing-scikit-learn: **adopted**

Misc
Rewrite "open-source" as "open source"
: **adopted**

(comments on highlighting CONUS and GCAM coupling to follow)
