+++ docs/source/tell.tests.rst
@@ -1,21 +1,0 @@
-tell.tests package
-==================
-
-Submodules
-----------
-
-tell.tests.test\_package\_data module
--------------------------------------
-
-.. automodule:: tell.tests.test_package_data
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: tell.tests
-   :members:
-   :undoc-members:
-   :show-inheritance:
