+++ .github/workflows/deploy.yml
@@ -2,7 +2,7 @@ name: Deploy
 
 on:
   push:
-    branches: [ review_edits ]
+    branches: [ main ]
 
 jobs:
   build:
