### 2022-08-05 IMMM-SFA/tell#36 @mcgrathc
Please see this PR for the changes:https://github.com/IMMM-SFA/tell/pull/41

Installation
tell is based on open-source publicly accessible data.
(https://immm-sfa.github.io/tell/tell_quickstarter.html#install-the-package-of-data-underpinning-tell)
The term "open data" rather than "open source" is more appropriate, cf. e.g. [this post](https://forum.openmod.org/t/definitions-for-open/1407): **adopted**

API reference
From the API reference on the website I was unable to find information on tell.train as mentioned in the quickstarter notebook. Maybe one or more modules are not properly included in the API reference?: **updated**

Spatial resolution
For temporal resolution the documentation states hourly resolution. For spatial resolution, the user has to read quite far to find:
This package is only inteded for the CONUS region, a subset of USA
What the adequate spatial resolution is ("Work at a spatial resolution adequate for input to a unit commitment/economic dispatch (UC/ED) model.").
I suggest both aspects, as they are properties of major relevance to interested potential users, should be placed as prominently as the information on the hourly resolution.: **added to quickstarter**

I also suggest you highlight the aspect of geographical scope and resolution in the beginning of the JOSS paper.:**The authors believe the first statement in the summary highlights both the geographical scope and resolution: _The purpose of the Total ELectricity Load (`tell`) model is to generate 21st century profiles of hourly electricity load (demand) across the Conterminous United States (CONUS)_.**

Target audience
Consider specififying the target audience and users of your package in a popular location like [here](https://immm-sfa.github.io/tell/index.html): **added to index**

comments on including MLP settings for user modification to follow
