### 2022-08-08 #4472 @fraukewiese
Thank you very much @robbiemorrison for your assessment regarding https://github.com/IMMM-SFA/tell/issues/31. Considering your thoughts and information about it, I think disclaimer and license of the software under consideration are ok for the publication in JOSS.
