### 2022-08-08 #4472 @mcgrathc
Thank you @robbiemorrison and @fraukewiese for your input on https://github.com/IMMM-SFA/tell/issues/31. I will close the issue now if we are in agreement that the disclaimer and license of the software under consideration are ok for the publication in JOSS.
