### 2022-08-08 IMMM-SFA/tell#31 @mcgrathc
Closing for now in agreement that the disclaimer and license of the software under consideration are ok for the publication in JOSS. Will open again if needed
