### 2022-08-17 611193b
add tests for mlp utils
