### 2022-08-17 00b93b6
Final issues
Minor updates to the paper and user guide in response to concerns from the reviewers.
