+++ paper/paper.md
@@ -34,7 +34,7 @@ affiliations:
     index: 1
   - name: Joint Global Change Research Institute, PNNL, College Park, MD, USA
     index: 2
-date: 03 June 2022
+date: 17 August 2022
 bibliography: paper.bib
 ---
 
@@ -52,7 +52,7 @@ United States (U.S.) state-level, annual projections from a national- to global-
 energy-economy model. This model accounts for a wide range of other factors affecting 
 electricity demand, including technology change in the building sector, energy prices, and 
 demand elasticities, which stems from model coupling with the U.S. version of the 
-Global Change Analysis Model (GCAM-USA).  `tell` was developed as part of
+Global Change Analysis Model (GCAM-USA). `tell` was developed as part of
 the Integrated Multisector Multiscale Modeling (IM3) project. IM3 explores the vulnerability and resilience of interacting energy,
 water, land, and urban systems in response to compound stressors, such as climate trends, 
 extreme events, population, urbanization, energy system transitions, and technology change.
