### 2022-08-18 IMMM-SFA/tell#36 @mcgrathc
**Below addresses the remaining comments on including MLP settings for user modification:**

**Comment:** The quickstarter mentions default settings for MLP training as stored in the GitHub
repository. I think these settings should be linked (to the respective file) or included in
the documentation. Consider linking the file or including the file contents in the
documentation.  It might be potentially interesting for users to look up and change these
settings in their local setup; especially if these settings might change between versions.
Consider describing the local availability of the settings file in the Python installation
location ( tell._file_ folder) so show the user how to check the values used by their
version.
Response: We thank the reviewer for making this suggestion. We’ve modified the MLP section of the
tell User Guide to describe where the default settings for the MLP models are stored in
the tell repository. We also explicitly link to the mlp_settings.yml file in the User Guide.
“Details of the MLP predictive variables are included in the table below. The default
parameter settings for training the MLP models are stored in the `mlp_settings.yml
<https://github.com/IMMM-SFA/tell/blob/main/tell/data/mlp_settings.yml>`_ file in
*/data* folder in the **tell** repository. The hyperparameters for the **tell** MLP
models (e.g., hidden layer sizes, maximum iterations, and validation fraction) were
determined using a grid search approach. Hyperparameters were allowed to vary across
BAs. Default hyperparameters for each BA are also included in the */data/models* folder
in the **tell** repository.”
