### 2022-08-18 11cdd1c
delete visualiztion test
