### 2022-08-18 IMMM-SFA/tell#35 @mcgrathc
We have increased code coverage via specifically for the MLP testing in `test_mlp_utils.py`
