### 2022-08-18 205b64d
Minor changes to paper and bibliography.
