+++ paper/paper.md
@@ -34,12 +34,12 @@ affiliations:
     index: 1
   - name: Joint Global Change Research Institute, PNNL, College Park, MD, USA
     index: 2
-date: 17 August 2022
+date: 18 August 2022
 bibliography: paper.bib
 ---
 
-[GitHub repoistory]: https://github.com/IMMM-SFA/tell 
-[Jupyter notebook]: https://immm-sfa.github.io/tell/tell_quickstarter.html
+[GitHub Repository]: https://github.com/IMMM-SFA/tell 
+[Jupyter Notebook]: https://immm-sfa.github.io/tell/tell_quickstarter.html
 [User Guide]: https://immm-sfa.github.io/tell/user_guide.html
 
 # Summary 
@@ -91,8 +91,8 @@ other regions with adequate historical data.
 `tell` integrates aspects of both short- and long-term projections of electricity demand in a coherent
 and scalable way. `tell` takes time series meteorological data at one-hour resolution as input and
 uses the temporal variations in weather to project hourly time-series of total electricity demand.
-The core predictions in `tell` are based on a series of multilayer perceptron (MLP) models that 
-relate historical meteorology to coincident BA-scale hourly loads for 54 independent BAs [@scikit-learn].
+The core predictions in `tell` are based on a series of multilayer perceptron (MLP) models [@scikit-learn] that 
+relate historical meteorology to coincident BA-scale hourly loads for 54 independent BAs.
 The BA load projections are first disaggregated to the county-level and then summed and scaled to 
 match the annual state-level total electricity demands projected by the U.S. version of the 
 Global Change Analysis Model (GCAM-USA) [@iyer2017; @binsted2022]. GCAM-USA is designed to 
@@ -107,7 +107,7 @@ The basic workflow for `tell` proceeds in six sequential steps:
 time-series of total electricity demand for 54 BAs that report their hourly loads in the EIA-930 
 dataset (Fig. 1a).
 2.	Use the empirical models to project future hourly loads for each BA based on IM3’s future 
-climate scenarios [@wrf].
+climate scenarios generated using the Weather Research and Forecasting [@wrf] model.
 3.	Distribute the hourly loads for each BA to the counties that BA operates in and then aggregate
  the county-level hourly loads from all BAs into annual state-level loads (Fig. 1b-g).
 4.	Calculate annual state-level scaling factors that force the bottom-up annual state-level total
@@ -124,8 +124,8 @@ for TELL; b) County-level populations within the ISNE BA service territory in 20
 total population within the BA that lives in each county; and d-g) Time-series of projected hourly 
 electricity demand in 2019 for select counties based on their population weights.*
 
-`tell` is an open source model that can be accessed via the [GitHub repoistory]. 
-The repository also includes a [Jupyter notebook] that provides a walkthrough of the core functionality 
+`tell` is an open source model that can be accessed via the [GitHub Repository]. 
+The repository also includes a [Jupyter Notebook] that provides a walk through of the core functionality 
 of `tell`. This notebook also provides easy access to statistical validation results between the forecasts 
 and observed data. Finally, more details about how the  model was formulated and its intended purpose 
 can be found in the `tell` [User Guide].
