+++ docs/source/index.rst
@@ -21,7 +21,7 @@ TELL documentation
 `tell` was created to:
 
   1. Project the short- and long-term evolution of hourly electricity demand in response to changes in weather and climate.
-  2. Work at a spatial resolution adequate for input to a unit commitment/economic dispatch (UC/ED) model (CONUS level).
+  2. Work at a spatial resolution adequate for input to a unit commitment/economic dispatch (UC/ED) model.
   3. Maintain consistency with the long-term growth and evolution of annual state-level electricity demand projected by an economically driven human-Earth system model.
 
 `tell` is targeted for users that are interested in grid stress modeling. This could include those whose business needs require
