### 2022-08-26 209cec4
add test data dir and execute forward tests
