+++ tell/tests/test_data_process_population.py
@@ -1,4 +1,5 @@
 import os
+import pkg_resources
 import unittest
 
 import tell.data_process_population as dpp
@@ -8,25 +9,21 @@ class TestProcessPop(unittest.TestCase):
     """Tests for functionality within execue forward.py"""
 
     # supporting data
-    TEST_POP_DIR = os.path.join(os.path.dirname(__file__), 'data/')
+    TEST_POP_DIR = pkg_resources.resource_filename("tell", os.path.join("tests", "data"))
 
     def test_fips_pop_yearly(self):
-        """Test to ensure high level functionality of normalize_prediction_data()"""
+        """Test to ensure high level functionality of fips_pop_yearly()"""
 
         # interpolate pop data from hourly to annual
         pop_df = dpp.fips_pop_yearly(pop_input_dir=TestProcessPop.TEST_POP_DIR,
                                      start_year=2000,
                                      end_year=2010)
 
-        # check that length is as expected
-        self.assertEqual(209, len(pop_df))
-
-        # check that length is as expected
-        self.assertEqual(3, pop_df.shape[1])
+        # check the shape of the data frame is expected
+        self.assertEqual((209, 3), pop_df.shape)
 
         # ensure all strings have a length greater than 0
-        self.assertEqual(True, all(len(i) > 0 for i in pop_df['county_fips']))
-
+        self.assertFalse(pop_df["county_FIPS"].isna().all())
 
 if __name__ == '__main__':
     unittest.main()
