+++ tell/install_forcing_data.py
@@ -20,7 +20,11 @@ class InstallForcingSample:
     # URL for DOI minted example data hosted on Zenodo
     DATA_VERSION_URLS = {'0.0.1': 'https://zenodo.org/record/6354665/files/sample_forcing_data.zip?download=1',
                          '0.1.0': 'https://zenodo.org/record/6354665/files/sample_forcing_data.zip?download=1',
-                         '0.1.1': 'https://zenodo.org/record/6354665/files/sample_forcing_data.zip?download=1'}
+                         '0.1.1': 'https://zenodo.org/record/6354665/files/sample_forcing_data.zip?download=1',
+                         '0.1.2': 'https://zenodo.org/record/6354665/files/sample_forcing_data.zip?download=1',
+                         '0.1.3': 'https://zenodo.org/record/6354665/files/sample_forcing_data.zip?download=1'}
+
+    DEFAULT_VERSION = 'https://zenodo.org/record/6354665/files/sample_forcing_data.zip?download=1'
 
     def __init__(self, data_dir=None):
         self.data_dir = data_dir
@@ -35,9 +39,11 @@ class InstallForcingSample:
             data_link = InstallForcingSample.DATA_VERSION_URLS[current_version]
 
         except KeyError:
-            msg = f"Link to data missing for current version: {current_version}. Please contact an administrator."
+            msg = f"Link to data missing for current version: {current_version}. Using default version: {InstallForcingSample.DEFAULT_VERSION}"
 
-            raise KeyError(msg)
+            data_link = InstallForcingSample.DEFAULT_VERSION
+
+            print(msg)
 
         # Retrieve content from the URL:
         print(f"Downloading the sample forcing data package for tell version {current_version}...")
