### 2022-08-29 #4472 @mcgrathc
> @fraukewiese, @euronion, @chrklemm. Thank you for all your helpful comments and suggestions. We have addressed all comments and made the appropriate changes to the package/paper that are available on the[ `main` branch.](https://github.com/IMMM-SFA/tell/tree/main) and are now ready to move forward with the review.

Please note we released a new version `v0.1.3` for some bug fixes. So please move forward with `v0.1.3`. Thank you
