### 2022-09-15 #4472 @chrklemm
Dear @mcgrathc,
many thanks for the extensive revision of the tell-package.

I would like to address two of my original criticisms again:

- Installation: During the installation under Windows, an error message still occurs: ![tell_installation_issue](https://user-images.githubusercontent.com/48713015/190387445-082e4b51-9263-45c5-8b88-a03e3a24fd7c.png)

- Test-coverage: Despite the new tests, test coverage has not increased significantly. According to codecov (https://app.codecov.io/gh/IMMM-SFA/tell), it is around 30 %. In the literature, a value of at least 60 % is recommended (https://doi.org/10.1371/journal.pcbi.1009481). However, I am not an expert on test coverage and would rely on the oppinion of the second reviewer @euronion if he agrees with the current state.

Best regards,
@cklm
