### 2022-09-26 #4472 @euronion
> Test-coverage: Despite the new tests, test coverage has not increased significantly. According to codecov (https://app.codecov.io/gh/IMMM-SFA/tell), it is around 30 %. In the literature, a value of at least 60 % is recommended (https://doi.org/10.1371/journal.pcbi.1009481). However, I am not an expert on test coverage and would rely on the oppinion of the second reviewer @euronion if he agrees with the current state.

I share @chrklemm opinion. The test coverage improved a little bit but is still low. Since there are a number of additional functionalities included in the package which do not represent the packages core functionality (e.g. visualisation) I would be fine with a test coverage in the range of 50% as long as core functionalities are covered by the tests.
Specifically the tests to "ensure high level functionality" (cf. https://github.com/IMMM-SFA/tell/issues/35) should be implemented to test the core functions of the package to a higher detail as well as the data processing parts which should ensure the correct data format and ideally data integrity.

I'm satisfied with everything else (except for the test coverage). @fraukewiese
