### 2022-10-03 134833d
test for denorm fn
