+++ tell/tests/test_mlp_utils.py
@@ -1,6 +1,7 @@
 import unittest
 
 import numpy as np
+import pandas as pd
 
 import tell.mlp_utils as mpu
 
@@ -27,6 +28,11 @@ class TestMlpUtils(unittest.TestCase):
                       'x_test_norm': np.array([[1.52,  0.3, -0.23],
                                               [1.61, 0.09, -0.36]]),
                       'y_test_norm': np.array([1.36, 0.5, 0.09])}
+
+    COMP_DENORM_DF = pd.DataFrame({"datetime": [1.1, 2.2],
+                                   "predictions": [3.4, 7.8],
+                                   "ground_truth": [1.1, 2.2],
+                                   "region": ["alpha"]*2})
 
     def test_normalize_prediction_data(self):
         """Test to ensure high level functionality of normalize_prediction_data()"""
@@ -61,6 +67,20 @@ class TestMlpUtils(unittest.TestCase):
         d = mpu.get_balancing_authority_to_model_dict()
 
         self.assertEqual(54, len(d))
+
+    def test_denormalize_features(self):
+        """Test to ensure high level functionality of denormalize_features()"""
+
+        norm_dict = {"max_y_train": np.array([3.1, 4.2]),
+                     "min_y_train": np.array([0.1, 1.2])}
+
+        df = mpu.denormalize_features(region="alpha",
+                                      normalized_dict=norm_dict,
+                                      y_predicted_normalized=np.array([1.1, 2.2]),
+                                      y_comparison=np.array([1.1, 2.2]),
+                                      datetime_arr=np.array([1.1, 2.2]))
+
+        pd.testing.assert_frame_equal(TestMlpUtils.COMP_DENORM_DF, df)
 
 
 if __name__ == '__main__':
