+++ tell/tests/test_mlp_predict.py
@@ -1,12 +1,26 @@
 import unittest
+import pkg_resources
+
+import pandas as pd
+
+import tell.mlp_predict as mp
 
 
-class TestPackageData(unittest.TestCase):
+class TestMlpPredict(unittest.TestCase):
     """Tests for functionality within mlp_predict.py"""
 
-    def test_mlp_predict(self):
-        """Test to ensure high level functionality of mlp_predict.py()"""
-        pass
+    COMP_PREDICT_DF = pd.read_parquet(pkg_resources.resource_filename("tell", "tests/data/comp_predict.parquet"))
+
+    def test_predict(self):
+        """Test to ensure high level functionality of predict()"""
+
+        df = mp.predict(region="ERCO",
+                        year=2039,
+                        data_dir=pkg_resources.resource_filename("tell", "tests/data"))
+
+        df.to_parquet(pkg_resources.resource_filename("tell", "tests/data/comp_predict.parquet"))
+
+        pd.testing.assert_frame_equal(TestMlpPredict.COMP_PREDICT_DF, df)
 
 
 if __name__ == '__main__':
