### 2022-10-03 d82041d
tests for batch predict
