+++ tell/install_raw_data.py
@@ -22,7 +22,8 @@ class InstallRawData:
                          '0.1.0': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1',
                          '0.1.1': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1',
                          '0.1.2': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1',
-                         '0.1.3': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1'}
+                         '0.1.3': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1',
+                         '0.1.4': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1'}
 
     DEFAULT_VERSION = 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1'
 
