### 2022-10-03 d4f2723
test for train mlp model
