### 2022-10-03 ab7185c
add train tests
