+++ MANIFEST.in
@@ -4,3 +4,4 @@ include tell/data/mlp_settings.yml
 include tell/data/hyperparameters.csv
 include tell/data/models/*
 include tell/tests/data/*
+include tell/tests/data/train_data/*
