### 2022-10-03 50b7fa8
tests for data install protocol
