### 2022-10-03 #4472 @mcgrathc
@fraukewiese, @euronion, @chrklemm, We thank the reviewers for raising the coverage issue. In response to this suggestion, we’ve again reviewed the codebase for opportunities to increase our code coverage. We implemented several new tests which brought the code coverage up to 48.49% for the package. All of the core functions of `tell` are now covered. The changes are available on the `main`[ branch](https://github.com/IMMM-SFA/tell) for `v0.1.4`

@chrklemm , You are encountering an issue common to many developers/modelers who utilize packages that have lower level dependencies.  They are especially tricky because the developer of higher-level packages like our own `tell` don't really have control over installation protocol of these third party dependencies per operating system.  What we can do, is help out by offering some insight from other popular, community-wide packages like GeoPandas (https://geopandas.org/en/stable/getting_started/install.html#installation) who are also working to make life easier on our geospatial community of users.  They recommend using a conda environment if you are so inclined.  You can still pip install `tell` into a conda environment as well.  Also, their developers recommend that you may need to compile GDAL and it's Python bindings directly via instructions from the source site.
