+++ tell/execute_forward.py
@@ -9,9 +9,11 @@ from pandas import DataFrame
 from scipy import interpolate
 from .states_fips_function import state_metadata_from_state_abbreviation
 
+def extract_gcam_usa_loads(scenario_to_process: str, filename: str) -> DataFrame:
+    """Extracts the state-level annual loads from a GCAM-USA output file.
 
-def extract_gcam_usa_loads(filename: str) -> DataFrame:
-    """Extracts the state-level annual loads from a given GCAM-USA output file.
+    :param scenario_to_process: Scenario to process
+    :type scenario_to_process:  str
 
     :param filename:            Name of the GCAM-USA output file
     :type filename:             str
@@ -22,6 +24,9 @@ def extract_gcam_usa_loads(filename: str) -> DataFrame:
 
     # Load in the raw GCAM-USA output file:
     gcam_usa_df = pd.read_csv(filename, index_col=None, header=0)
+
+    # Subset the data to only the scenario you want to process:
+    gcam_usa_df = gcam_usa_df[gcam_usa_df['scenario'].isin([scenario_to_process])]
 
     # Subset the data to only the total annual consumption of electricity by state:
     gcam_usa_df = gcam_usa_df[gcam_usa_df['param'].isin(['elecFinalBySecTWh'])]
@@ -39,8 +44,8 @@ def extract_gcam_usa_loads(filename: str) -> DataFrame:
         (state_fips, state_name) = state_metadata_from_state_abbreviation(states[i])
 
         # Linearly interpolate the 5-year loads from GCAM-USA to an annual time step:
-        annual_time_vector = pd.Series(range(subset_df['origX'].min(), subset_df['origX'].max()))
-        interpolation_function = interpolate.interp1d(subset_df['origX'], subset_df['value'], kind='linear')
+        annual_time_vector = pd.Series(range(subset_df['x'].min(), subset_df['x'].max()))
+        interpolation_function = interpolate.interp1d(subset_df['x'], subset_df['value'], kind='linear')
         annual_loads = interpolation_function(annual_time_vector)
 
         # Create an empty dataframe and store the results:
@@ -170,7 +175,7 @@ def aggregate_mlp_output_files(list_of_files: list) -> DataFrame:
     return mlp_output_df
 
 
-def output_tell_summary_data(joint_mlp_df: DataFrame, year_to_process: str, data_output_dir: str):
+def output_tell_summary_data(joint_mlp_df: DataFrame, year_to_process: str, gcam_target_year: str, data_output_dir: str):
     """Writes a summary file describing state-level annual total loads from TELL and GCAM-USA.
 
     :param joint_mlp_df:            DataFrame of processed TELL loads
@@ -178,6 +183,9 @@ def output_tell_summary_data(joint_mlp_df: DataFrame, year_to_process: str, data
 
     :param year_to_process:         Year to process
     :type year_to_process:          str
+
+    :param gcam_target_year:        Year to scale against the GCAM-USA annual loads
+    :type gcam_target_year:         str
 
     :param data_output_dir:         Data output directory
     :type data_output_dir:          str
@@ -218,7 +226,8 @@ def output_tell_summary_data(joint_mlp_df: DataFrame, year_to_process: str, data
     output_df = output_df.sort_values('State_Name')
 
     # Generate the .csv output file name:
-    csv_output_filename = os.path.join(data_output_dir, 'TELL_State_Summary_Data_' + year_to_process + '.csv')
+    csv_output_filename = os.path.join(data_output_dir, 'TELL_State_Summary_Data_' + year_to_process
+                                       + '_Scaled_' + gcam_target_year + '.csv')
 
     # Write out the dataframe to a .csv file:
[…]
