### 2022-10-13 1621f5e
Added granularity to directory structure
