### 2022-10-18 1a9ec43
add geopandas dependencies statements and update version
