+++ README.md
@@ -17,7 +17,8 @@
 ### Install `tell`
 
 `tell` is available via GitHub repository by using the pip install functionality. `tell` requires a Python version between 3.8 and 4.0 as well as a pip install to import the package. `tell` has been tested on
-Windows and Mac platforms.
+Windows and Mac platforms. (Note: For those installing on Windows, `tell`  is supported by GeoPandas functionality.  Please see suggestions for installing GeoPandas on Windows here:  
+https://geopandas.org/en/stable/getting_started/install.html)
 
 ```bash
 pip install tell
