### 2022-10-18 #4472 @mcgrathc
Thank you @euronion and @chrklemm for your thoughtful review!

@chrklemm we have added instructions for the Windows installation in the newest release `v0.1.5`

@fraukewiese thank you for your help in this process!

**We appreciate everyone's time and effort committed to improving this package!**
