+++ tell/install_raw_data.py
@@ -24,7 +24,8 @@ class InstallRawData:
                          '0.1.2': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1',
                          '0.1.3': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1',
                          '0.1.4': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1',
-                         '0.1.5': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1'}
+                         '0.1.5': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1',
+                         '1.0.0': 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1'}
 
     DEFAULT_VERSION = 'https://zenodo.org/record/6378036/files/tell_raw_data.zip?download=1'
 
