### 2022-10-24 #4472 @mcgrathc
@fraukewiese the new version tag is `v1.0.0`

The reviewed software has been archived with the updated metadata here: 10.5281/zenodo.7246116

Thank you for your help in moving this forward.
