+++ production_visualizations/README.md
@@ -1,0 +1,15 @@
+# Quickview Analysis of TELL Load Duration Curve Variability
+>
+## Purpose
+TBD
+>
+## Notes
+>
+TBD
+>
+## Balancing Authority Analyses
+>
+| Name | rcp85hotter_ssp5 |
+| :-: | :-: |
+| AEC | [AEC LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AEC_rcp85hotter_ssp5.png) |
+
