### 2022-11-02 f992535
Added lots more BAs
