+++ production_visualizations/README.md
@@ -1,15 +1,45 @@
 # Quickview Analysis of TELL Load Duration Curve Variability
 >
 ## Purpose
-TBD
+This analysis looks at the interannual variability in the load duration curve (LDC) for each Balancing Authority (BA) 
+and state simulated by the TELL model.
 >
 ## Notes
 >
-TBD
+1. Right now the plots are set to focus on the highest 168 hours (7 days) of load for each BA and state. This helps to
+demonstrate variability in peak loads.
 >
 ## Balancing Authority Analyses
 >
 | Name | rcp85hotter_ssp5 |
 | :-: | :-: |
+| AEC | [AEC LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AEC_rcp85hotter_ssp5.png) |
+| AECI | [AECI LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AECI_rcp85hotter_ssp5.png) |
+| AVA | [AVA LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AVA_rcp85hotter_ssp5.png) |
+| AVRN | [AVRN LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AVRN_rcp85hotter_ssp5.png) |
+| AZPS | [AZPS LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AZPS_rcp85hotter_ssp5.png) |
+| BANC | [BANC LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_BANC_rcp85hotter_ssp5.png) |
+| BPAT | [BPAT LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_BPAT_rcp85hotter_ssp5.png) |
+| CHPD | [CHPD LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_CHPD_rcp85hotter_ssp5.png) |
+| CISO | [CISO LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_CISO_rcp85hotter_ssp5.png) |
+| CPLE | [CPLE LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_CPLE_rcp85hotter_ssp5.png) |
+| CPLW | [CPLW LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_CPLW_rcp85hotter_ssp5.png) |
+| DEAA | [DEAA LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_DEAA_rcp85hotter_ssp5.png) |
+| DOPD | [DOPD LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_DOPD_rcp85hotter_ssp5.png) |
+| DUK | [DUK LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_DUK_rcp85hotter_ssp5.png) |
+| EEI | [EEI LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_EEI_rcp85hotter_ssp5.png) |
+| EPE | [EPE LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_EPE_rcp85hotter_ssp5.png) |
+| ERCO | [ERCO LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_ERCO_rcp85hotter_ssp5.png) |
+| FMPP | [FMPP LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_FMPP_rcp85hotter_ssp5.png) |
[…]
