### 2022-11-02 bb4e290
Simplified BAs
