+++ production_visualizations/README.md
@@ -11,35 +11,9 @@ demonstrate variability in peak loads.
 >
 ## Balancing Authority Analyses
 >
-| Name | rcp85hotter_ssp5 |
-| :-: | :-: |
-| AEC | [AEC LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AEC_rcp85hotter_ssp5.png) |
-| AECI | [AECI LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AECI_rcp85hotter_ssp5.png) |
-| AVA | [AVA LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AVA_rcp85hotter_ssp5.png) |
-| AVRN | [AVRN LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AVRN_rcp85hotter_ssp5.png) |
-| AZPS | [AZPS LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AZPS_rcp85hotter_ssp5.png) |
-| BANC | [BANC LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_BANC_rcp85hotter_ssp5.png) |
-| BPAT | [BPAT LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_BPAT_rcp85hotter_ssp5.png) |
-| CHPD | [CHPD LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_CHPD_rcp85hotter_ssp5.png) |
-| CISO | [CISO LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_CISO_rcp85hotter_ssp5.png) |
-| CPLE | [CPLE LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_CPLE_rcp85hotter_ssp5.png) |
-| CPLW | [CPLW LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_CPLW_rcp85hotter_ssp5.png) |
-| DEAA | [DEAA LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_DEAA_rcp85hotter_ssp5.png) |
-| DOPD | [DOPD LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_DOPD_rcp85hotter_ssp5.png) |
-| DUK | [DUK LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_DUK_rcp85hotter_ssp5.png) |
-| EEI | [EEI LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_EEI_rcp85hotter_ssp5.png) |
-| EPE | [EPE LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_EPE_rcp85hotter_ssp5.png) |
-| ERCO | [ERCO LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_ERCO_rcp85hotter_ssp5.png) |
-| FMPP | [FMPP LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_FMPP_rcp85hotter_ssp5.png) |
-| AEC | [AEC LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AEC_rcp85hotter_ssp5.png) |
-| AEC | [AEC LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AEC_rcp85hotter_ssp5.png) |
-| AEC | [AEC LDC](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AEC_rcp85hotter_ssp5.png) |
[…]
