### 2022-11-02 74dd4a7
Shrunk figures
