+++ production_visualizations/README.md
@@ -11,9 +11,23 @@ demonstrate variability in peak loads.
 >
 ## Balancing Authority Analyses
 >
-| Name | rcp45cooler_ssp3 | rcp45cooler_ssp5 | rcp45hotter_ssp3 | rcp45hotter_ssp5 | rcp85cooler_ssp3 | rcp85cooler_ssp5 | rcp85hotter_ssp3 | rcp85hotter_ssp5 |
-| :-: | :-: | :-: | :-: | :-: | :-: | :-: | :-: | :-: |
-| CISO | - | - | - | - | - | - | - | [Plot](rcp85hotter_ssp5/TELL_BA_LDC_Variability_CISO_rcp85hotter_ssp5.png) |
-
-
+| Name | rcp85cooler_ssp3 | rcp85cooler_ssp5 | rcp85hotter_ssp3 | rcp85hotter_ssp5 |
+| :-: | :-: | :-: | :-: | :-: |
+| CISO | - | - | - | [Plot](rcp85hotter_ssp5/TELL_BA_LDC_Variability_CISO_rcp85hotter_ssp5.png) |
+| PJM | - | - | - | [Plot](rcp85hotter_ssp5/TELL_BA_LDC_Variability_PJM_rcp85hotter_ssp5.png) |
+| ERCO | - | - | - | [Plot](rcp85hotter_ssp5/TELL_BA_LDC_Variability_ERCO_rcp85hotter_ssp5.png) |
+| ISNE | - | - | - | [Plot](rcp85hotter_ssp5/TELL_BA_LDC_Variability_ISNE_rcp85hotter_ssp5.png) |
+| MISO | - | - | - | [Plot](rcp85hotter_ssp5/TELL_BA_LDC_Variability_MISO_rcp85hotter_ssp5.png) |
+| AZPS | - | - | - | [Plot](rcp85hotter_ssp5/TELL_BA_LDC_Variability_AZPS_rcp85hotter_ssp5.png) |
+>
+## State Analyses
+>
+| Name | rcp85cooler_ssp3 | rcp85cooler_ssp5 | rcp85hotter_ssp3 | rcp85hotter_ssp5 |
+| :-: | :-: | :-: | :-: | :-: |
+| California | - | - | - | [Plot](rcp85hotter_ssp5/TELL_State_LDC_Variability_California_rcp85hotter_ssp5.png) |
+| Texas | - | - | - | [Plot](rcp85hotter_ssp5/TELL_State_LDC_Variability_Texas_rcp85hotter_ssp5.png) |
+| New York | - | - | - | [Plot](rcp85hotter_ssp5/TELL_State_LDC_Variability_New_York_rcp85hotter_ssp5.png) |
+| Pennsylvania | - | - | - | [Plot](rcp85hotter_ssp5/TELL_State_LDC_Variability_Pennsylvania_rcp85hotter_ssp5.png) |
+| Florida | - | - | - | [Plot](rcp85hotter_ssp5/TELL_State_LDC_Variability_Florida_rcp85hotter_ssp5.png) |
+| Washington | - | - | - | [Plot](rcp85hotter_ssp5/TELL_State_LDC_Variability_Washington_rcp85hotter_ssp5.png) |
 
