### 2022-11-04 e5bd6c7
Added figure
