+++ production_visualizations/README_LDCs.md
@@ -12,8 +12,8 @@ demonstrate variability in peak loads.
 ## Summary Figures
 The figures below show box-plots of the mean difference between annual load profiles (e.g., each year from 2020-2030 
 ) and the center year profile (e.g., 2025) for all states and BAs. The top panel shows the distribution of the actual 
-load difference in MWh and the bottom panel shows the distribution of the relative difference (i.e., [Year - 
-Center Year] /  Center Year). The calculations are for the top 168 hours of load for each individual state or BA.
+load difference in MWh and the bottom panel shows the distribution of the relative difference (i.e., abs[Center Year - Year] 
+/  Center Year). The calculations are for the top 168 hours of load for each individual state or BA.
 ![State_Peak_Load_Bias_Plot](rcp85hotter_ssp5/Peak_Load_Variability_States_Top_168_rcp85hotter_ssp5.png)
 ![BA_Peak_Load_Bias_Plot](rcp85hotter_ssp5/Peak_Load_Variability_BAs_Top_168_rcp85hotter_ssp5.png)
 
