### 2022-11-10 #4472 @kyleniemeyer
Hi @mcgrathc, I made some edits for the paper's bibliography (author names need to be separated by `and` rather than a comma), could you merge my PR? https://github.com/IMMM-SFA/tell/pull/44
