### 2022-11-10 #4472 @kyleniemeyer
@mcgrathc sorry, looks like I accidentally added some duplicate "and"s before final author names. Fixed in https://github.com/IMMM-SFA/tell/pull/45 and then we should be good to go
