+++ paper/paper.bib
@@ -131,7 +131,7 @@
 @misc{wrf,
  title={A Description of the Advanced Research {WRF} Version 4.1},
  author={Skamarock, W. C. and Klemp, J. B. and Dudhia, J. and Gill, D. O. and Liu, Z. and Berner, J. and
-        Wang, W. and Powers, J. G. and Duda, M. G. and Barker, D. M. and Huang X.-Y.},
+        Wang, W. and Powers, J. G. and Duda, M. G. and Barker, D. M. and Huang, X.-Y.},
  howpublished={NCAR Technical Note NCAR/TN-556+STR},
  year={2019},
  doi={10.5065/1dfh-6p97}
