### 2023-04-10 #4472 @mcgrathc
Hi @kyleniemeyer, it has come to my attention that there is an issue with the funding statement. It includes the words "in-part" for the funding source when it was the total funding source. Is there a way to remove the words "in-part" after the paper has been published? Thank you for your time!

The statement below is where the "in-part" should be removed:
_This research was supported **in part** by the U.S. Department of Energy, Office of Science, as part of research in MultiSector Dynamics, Earth and Environmental System Modeling Program. The Pacific Northwest National Laboratory is operated for DOE by Battelle Memorial Institute under contract DE-AC05-76RL01830. The views and opinions expressed in this paper are those of the authors alone_
